First :) Adding base project using neutrino react preset
This is using the beta of neutrino 9. I really wanted to use neutrino but the fkn stable build was broken out of the box due to dependency issues.
Committing all the node_modules at the moment for simplicity for other contributors.
If it becomes unwieldy or creates conflicts for other developer environments they can be removed.